Improvements for the MPSS daemon ebuild.
The ebuild is improved in two ways. First, the init.d script was completely rewritten. It uses now the openrc-run infrastructure. Furthermore, the init script now uses the module loading code supplied by intel. Therefore, a manual module loading is no longer necessary.

The second improvement targets the dependencies. The mpss-daemon ebuild not pulls in everything necessary to run the Xeon Phi. That includes both the kernel module and the kernel and userspace running on the Xeon Phi. These both dependencies are triggered by a use flag that is set by default. The intention in having a use flag for the kernel module is that Intel might integrate the driver into the main line kernel later on and therefore make the module obsolete for newer kernels. The intention of the firmware use flag is that the user might want to have a custom kernel and user land running on the Xeon Phi.