Add support for the ssh1 use flag to the mpss-daemon ebuild.
By default, the gentoo ssh configuration does not support rsa1 cyphers, which is required by the micmgmt tool. This ebuild adds support for the ssh1 use flag to handle this issue. If not set the support for the rsa1 cypher is removed from micmgmt. If set, the ebuild the ebuild requests ssh1 from openssh in order for micmgmt to work properly.